🔧 chore: Move UserAvatar
